requirements/requirements_pip.txt
amqp==2.6.1
asgiref==3.8.1
async-timeout==4.0.3
backports-datetime-fromisoformat==2.0.1
billiard==3.6.4.0
celery==4.4.7
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
cinemagoer==2023.5.1
click==8.1.7
colorama==0.4.6
cron-descriptor==1.4.3
cryptography==42.0.5
deezer-py==1.3.7
Django==4.2.11
django-appconf==1.0.6
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-imagekit==5.0.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-redis==5.4.0
django-shortcuts==1.6
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filetype==1.2.0
greenlet==3.0.3
gunicorn==21.2.0
idna==3.7
inflection==0.5.1
kombu==4.6.11
lxml==5.2.1
m3u8==4.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
mutagen==1.47.0
packaging==24.0
pilkit==3.0
pillow==10.3.0
prompt-toolkit==1.0.14
psycopg2-binary=2.9.9
psycopg2==2.9.9
py-deezer==1.1.4.post1
pycparser==2.22
pygments==2.17.2
PyInquirer==1.0.3
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
regex==2023.12.25
requests==2.31.0
rich==13.7.1
six==1.16.0
spotipy==2.23.0
SQLAlchemy==2.0.29
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==1.3.0
wcwidth==0.2.13
